chore, refactor: added access_token_life and refresh_token_life env. variables and changed the file structure
