+++ server/routes/index.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import path = require("path");
 
 const main = (req: Request, res: Response) => {
-  return res.sendFile(path.join(__dirname, "../public/index.html"));
+  return res.sendFile(path.join(__dirname, "public/index.html"));
 };
 
 export default main;
